TIP-842: Move attribute normalizer to Akeneo\Pim\Structure
