+++ src/Akeneo/Pim/Structure/Component/spec/Normalizer/ExternalApi/AttributeNormalizerSpec.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace spec\Akeneo\Tool\Component\Api\Normalizer;
+namespace spec\Akeneo\Pim\Structure\Component\Normalizer\ExternalApi;
 
 use PhpSpec\ObjectBehavior;
-use Akeneo\Tool\Component\Api\Normalizer\AttributeNormalizer;
+use Akeneo\Pim\Structure\Component\Normalizer\ExternalApi\AttributeNormalizer;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Prophecy\Argument;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
