+++ src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/normalizers.yml
@@ -6,7 +6,6 @@ parameters:
     pim_api.normalizer.family.class: Akeneo\Tool\Component\Api\Normalizer\FamilyNormalizer
     pim_api.normalizer.family_variant.class: Akeneo\Tool\Component\Api\Normalizer\FamilyVariantNormalizer
     pim_api.normalizer.category.class: Akeneo\Tool\Component\Api\Normalizer\CategoryNormalizer
-    pim_api.normalizer.attribute.class: Akeneo\Tool\Component\Api\Normalizer\AttributeNormalizer
     pim_api.normalizer.attribute_option.class: Akeneo\Tool\Component\Api\Normalizer\AttributeOptionNormalizer
     pim_api.normalizer.attribute_group.class: Akeneo\Tool\Component\Api\Normalizer\AttributeGroupNormalizer
     pim_api.normalizer.file.class: Akeneo\Tool\Component\Api\Normalizer\FileNormalizer
@@ -68,13 +67,6 @@ services:
         class: '%pim_api.normalizer.category.class%'
         arguments:
             - '@pim_catalog.normalizer.standard.category'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_api.normalizer.attribute:
-        class: '%pim_api.normalizer.attribute.class%'
-        arguments:
-            - '@pim_catalog.normalizer.standard.attribute'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
