+++ src/Pim/Bundle/EnrichBundle/Resources/config/normalizers.yml
@@ -13,7 +13,6 @@ parameters:
     pim_enrich.normalizer.group_type.class:                        Pim\Bundle\EnrichBundle\Normalizer\GroupTypeNormalizer
     pim_enrich.normalizer.family.class:                            Pim\Bundle\EnrichBundle\Normalizer\FamilyNormalizer
     pim_enrich.normalizer.family_variant.class:                    Pim\Bundle\EnrichBundle\Normalizer\FamilyVariantNormalizer
-    pim_enrich.normalizer.attribute.class:                         Pim\Bundle\EnrichBundle\Normalizer\AttributeNormalizer
     pim_enrich.normalizer.attribute.versioned.class:               Pim\Bundle\EnrichBundle\Normalizer\VersionedAttributeNormalizer
     pim_enrich.normalizer.group.class:                             Pim\Bundle\EnrichBundle\Normalizer\GroupNormalizer
     pim_enrich.normalizer.file.class:                              Pim\Bundle\EnrichBundle\Normalizer\FileNormalizer
@@ -170,15 +169,6 @@ services:
             - '@pim_enrich.provider.structure_version.group_type'
         tags:
             - { name: pim_internal_api_serializer.normalizer }
-
-    pim_enrich.normalizer.attribute:
-        class: '%pim_enrich.normalizer.attribute.class%'
-        arguments:
-            - '@pim_serializer'
-            - '@pim_enrich.provider.field.chained'
-            - '@pim_enrich.provider.empty_value.chained'
-            - '@pim_enrich.provider.filter.chained'
-            - '@pim_catalog.localization.localizer.number'
 
     pim_enrich.normalizer.entity_with_family_variant:
         class: '%pim_enrich.normalizer.entity_with_family_variant.class%'
